feat: 예약이 존재할 경우 시간 삭제 불가 기능 추가

+++ src/main/java/roomescape/domain/reservationtime/ReservationTimeService.java
@@ -3,6 +3,7 @@ package roomescape.domain.reservationtime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.reservationtime.dto.CreateTimeRequest;
 import roomescape.domain.reservationtime.dto.CreateTimeResponse;
 import roomescape.domain.reservationtime.dto.ReservationTimeResponse;
@@ -12,6 +13,7 @@ import roomescape.domain.reservationtime.dto.ReservationTimeResponse;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
     public CreateTimeResponse createReservationTime(CreateTimeRequest request) {
         request.validate();
@@ -26,6 +28,9 @@ public class ReservationTimeService {
     }
 
     public void deleteReservation(Long id) {
+        if (reservationRepository.countByTimeId(id) > 0) {
+            throw new IllegalArgumentException("이미 예약이 존재할 경우 시간대를 삭제할 수 없습니다.");
+        }
         int deletedCount = reservationTimeRepository.deleteById(id);
         if (deletedCount == 0) {
             throw new IllegalArgumentException("존재하지 않는 예약 시간대 입니다.");
